refactor(cli): 拆分 ts_bridge JSON-RPC 与进程启动层
- 将 ts_bridge.rs 迁入 ts_bridge/ 模块目录
- 提取 constants 与 json_rpc 子模块（spawn/framing/process）
- 对外 pub use 路径保持不变
