+++ apps/cli/src/ts_bridge/constants.rs
@@ -1,0 +1,6 @@
+pub(crate) const ENV_RUNTIME_BACKEND_NODE_PATH: &str = "SPIRIT_NODE_PATH";
+pub(crate) const ENV_RUNTIME_BRIDGE_PATH: &str = "SPIRIT_AGENT_CORE_BRIDGE_PATH";
+pub(crate) const ENV_RUNTIME_HOST_INTERNAL_MODULE_PATH: &str = "SPIRIT_HOST_INTERNAL_MODULE_PATH";
+pub(crate) const ENV_RUNTIME_HOST_INTERNAL_SPIRIT_DATA_DIR: &str = "SPIRIT_HOST_INTERNAL_SPIRIT_DATA_DIR";
+pub(crate) const ENV_API_BASE: &str = "SPIRIT_API_BASE";
+pub(crate) const ENV_API_KEY: &str = "SPIRIT_API_KEY";
